pyproject.toml
[project]
name = "alpemones"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymongo>=4.15.3",
    "python-dotenv>=1.2.1",
]
